docs(ADR): Adds device naming convention
As we are only now adding our ADRs, we will be adding them
incrementally.

This ADR focuses on how we name our devices, this convention is crucial
to n-vyro.io as it is not only the name of the device, but it is also
the identifier used to associate a device with it related data.

As this documentation is quite extensive, there's no need to go into the
details here.

This commit adds the initial ADR, along with adds the `adr-tool`.

`adr-tool` is the script we'll be using to generate, and manage our ADRs
going forward.

Signed-off-by: Yomi Colledge <baphled@boodah.net>