chore(templates): Add issue templates
This will serve as our templates for creating issues and features
throughout our projects.

As it stands we've mostly been using our commit messages, and making
sure that they have enough information in them to then use in our PRs.

This is cool, but as we gather more contributors we're going to need to
standardise this in a way where there's limited interaction from other.

These templates are form-based, so they should help guide our
contributors through the steps of each process.

Signed-off-by: Yomi Colledge <baphled@boodah.net>