chore(ADR): Messaging topics
Now that we're about to address our missing alerts and notifications
topic, we need to take some time to docuement our reasoning behind our
naming convention and how we see this playing out for future devices.

Here we discuss what we've already got, and what we think we need to
implement, especially from the perspective of our notifications and
alerts messages.

We will also touch on the topic prefixes we think we should have, and
the reasoning behind each choice.

Signed-off-by: Yomi Colledge <baphled@boodah.net>