chore(ADR): Settings payload
We need to document how we handle changes to a device setting. On the
surface of things this kind of change seems simple. However, we need to
be able to listen out for these changes so that we can update our
ecosystem if a change impacts any user input.

We've yet to create the service for handling these changes, but we have
already implemented the functionality to publish the message.

Here we will go over the fact that this is the case, and we will discuss
what the payload looks like and, at a high level, how we envisage this
future service working.

Signed-off-by: Yomi Colledge <baphled@boodah.net>